Success load total_profit,toral_expense,total_sale. (+9 squashed commits) Squashed commits: [95c72b7] temp10 [438a6e0] temp8 [dc7efde] temp7 [19b0823] temp6 [b2ef9c5] temp5 [d8d392b] temp4 [fe6e8f1] temp3 [2764956] temp2 [24f6986] temp1

+++ src/controller/SalesDetailCRUDController.java
@@ -113,7 +113,7 @@ public class SalesDetailCRUDController implements Initializable {
 				sales_id, //ここに画面遷移前から引っ張ってきた売上IDを移す。
 				Long.valueOf(fx_combo_customers_id.getValue().substring(0,fx_combo_customers_id.getValue().indexOf(":"))),    //ここにコンボボックスのテキストを整形して仕入れ先IDを取り出す。
 				fx_text_detail.getText(),
-				Float.valueOf(fx_text_price.getText())
+				Long.valueOf(fx_text_price.getText())
 				);
 		new SalesDao(sql);
     	for ( int i = 0; i<fx_table.getItems().size(); i++) {
