+++ src/sql_crud/SalesDetails_FindAll.java
@@ -44,7 +44,7 @@ public class SalesDetails_FindAll implements ISQLExecutable {
 					Long sales_id = rs.getLong("sales_id");
 					Long vendor_id = rs.getLong("vendor_id");
 					String description = rs.getString("description");
-					Float price = rs.getFloat("price");
+					Long price = rs.getLong("price");
 					String customer_name = rs.getString("name");
 					
 					SalesDetails record = new SalesDetails(id,sales_id,vendor_id,description,price,customer_name);
